Fix handling of global and nonlocal keywords
The `global` keyword doesn't introduce a global variable, it only changes
the resolution rule for current scope.
The `nonlocal` keyword has a similar behavior.

Also pre-compute the set of local identifiers for each scope, and use this to
correctly report identifier that may be unbound at runtime. This should
fix #44