Do not clean lookup list too early
The name lookup is done in two steps: first we build the lookup list,
second we perform the lookup, walking the list.

While building the lookup list, we may find an invalid situation that we
need to report. Yet this situation may not happen if the lookup is
successful before we reach the invalid situation.

Take this scenario into account while performing the lookup.

Fix #57