Simplify visit_Name and fix annotations definitions linking to deleted names
Bugfix: If a deferred annotation refers to a deleted name, the declaration of name will be used as the defintion of the annotation. This is implemented by resolving the deferred annotations after the deleted nodes have been removed from the locals and falling back to killed definitions when none are live.

Refactoring:
- Factorize code related to locals addition in the method add_to_locals(),  set_or_extend_global() and extend_global() methods. These three methods are noe the only ones modifying the locals mapping.
- Change the type of locals from defaultdict(list) to defaultdict(ordered_set) for performance and clarty reasons.
- Cleanup legacy Python2 code as a drive-by change.