Improve support for imports
Relative imports will be parsed as expected only if beniget is given the module name explicitly when calling DefUseChains(modname="stuff.bar"). If the analyzed module is the __init__.py file of a packages, the argument is_package=True should also be passed to have correct understanding of relative imports.