Support PEP695 (Python 3.12)Type Parameters and PEP696 (Python3.13) Type Defaults
The strategy is to use a special node representing the annotation scopes and pass it to ScopeContext() before continuing the processing.

The compute_defs() and lookup_annotation_name_defs() method needed a small adjustment to correctly model the new scoping rules.

The code arround deferred annotations has been enhanced such that it validates that the annotation does not use an illegal syntax.

Fixes #82